% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{run_checkerboard_app}
\alias{run_checkerboard_app}
\title{Run the Checkerboard Shiny app}
\usage{
run_checkerboard_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\value{
A Shiny application object, invisibly.
}
\description{
Launch an interactive interface for importing checkerboard assay data,
assigning columns to \code{DrugA}, \code{DrugB}, optional \code{DrugC}, and \code{OD}, and
running the methods provided by the \link{bliss} R6 class.
}
